stuff
ManufactureInfoState
ResearchInfoState + make functions private
	- CTRL to increase/decrease by 10
CraftErrorState
ItemsArrivingState + init,SetPalette() in case player goes to Base
LowFuelState
ProductionCompleteState
UfoDetectedState
	- add init() to check if should show btn5Secs
GeoscapeState: new function is5Sec()

format & varname changes & cast

DAWNURBAN12.map, unblock door
DAWNDECOR.mcd, proper footstep sound on carpet

+++ src/Ruleset/Armor.cpp
@@ -116,7 +116,7 @@ void Armor::load(const YAML::Node& node)
 		for (size_t
 				i = 0;
 				i < dmg.size()
-					&& i < DAMAGE_TYPES;
+					&& i < static_cast<size_t>(DAMAGE_TYPES);
 				++i)
 		{
 			_damageModifier[i] = dmg[i]					.as<float>();
@@ -133,7 +133,7 @@ void Armor::load(const YAML::Node& node)
 
 	_forcedTorso = (ForcedTorso)node["forcedTorso"]		.as<int>(_forcedTorso);
 
-	if (_drawingRoutine == 0
+	if (   _drawingRoutine == 0
 		|| _drawingRoutine == 1
 		|| _drawingRoutine == 4
 		|| _drawingRoutine == 6
